requirements.txt
anthropic
openai
qdrant-client
watchdog
pdfplumber
python-docx
openpyxl
slack-bolt
python-dotenv
mcp
aiohttp
aiosqlite
python-pptx
beautifulsoup4
msal
msgraph-sdk
